@@ -8,14 +8,14 @@ description: >-
 
 homepage: https://github.com/dart-lang/intl_translation
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  analyzer: '>=0.39.12 <0.40.0'
-  args: '>=0.12.1 <2.0.0'
-  dart_style: ^1.0.0
-  intl: '>=0.15.3 <0.17.0'
-  path: '>=0.9.0 <2.0.0'
-  petitparser: ^3.0.0
+  analyzer: ^1.5.0
+  args: '^2.1.0'
+  dart_style: ^2.0.0
+  intl: '^0.17.0'
+  path: ^1.8.0
+  petitparser: ^4.1.0
 dev_dependencies:
-  test: ^1.2.0
+  test: ^1.17.1
