@@ -45,20 +45,20 @@ main(List<String> args) {
 String rewriteMessages(String source, String sourceName) {
   var messages = findMessages(source, sourceName);
   messages.sort((a, b) => a.sourcePosition.compareTo(b.sourcePosition));
-  var start = 0;
+  int? start = 0;
   var newSource = new StringBuffer();
   for (var message in messages) {
     if (message.examples != null) {
-      newSource.write(source.substring(start, message.sourcePosition));
+      newSource.write(source.substring(start!, message.sourcePosition));
       rewrite(newSource, source, start, message);
       start = message.endPosition;
     }
   }
-  newSource.write(source.substring(start));
+  newSource.write(source.substring(start!));
   return newSource.toString();
 }
 
-void rewrite(StringBuffer newSource, String source, int start, message) {
+void rewrite(StringBuffer newSource, String source, int? start, message) {
   var originalSource =
       source.substring(message.sourcePosition, message.endPosition);
   var examples = nonConstExamples.firstMatch(originalSource);
@@ -66,7 +66,7 @@ void rewrite(StringBuffer newSource, String source, int start, message) {
     newSource.write(originalSource);
   } else {
     var modifiedSource = originalSource.replaceFirst(
-        examples.group(1), examples.group(1) + 'const');
+        examples.group(1)!, examples.group(1)! + 'const');
     newSource.write(modifiedSource);
   }
 }
