@@ -8,7 +8,7 @@ import 'package:path/path.dart' as path;
 
 /// Takes a file with a list of file paths, one per line, and returns the names
 /// as paths in terms of the directory containing [fileName].
-Iterable<String> linesFromFile(String fileName) {
+Iterable<String> linesFromFile(String? fileName) {
   if (fileName == null) {
     return [];
   }
