@@ -22,8 +22,8 @@ const List<String> defaultFiles = const [
 ];
 
 void runTestWithWarnings(
-    {bool warningsAreErrors,
-    int expectedExitCode,
+    {required bool warningsAreErrors,
+    int? expectedExitCode,
     bool embeddedPlurals: true,
     List<String> sourceFiles: defaultFiles}) {
   verify(ProcessResult result) {
@@ -44,7 +44,7 @@ void runTestWithWarnings(
     args.add('--no-embedded-plurals');
   }
   var files = sourceFiles.map(asTempDirPath).toList();
-  List<String> allArgs = [program]..addAll(args)..addAll(files);
+  List<String?> allArgs = [program]..addAll(args)..addAll(files);
   var callback = expectAsync1(verify);
 
   run(null, allArgs).then(callback);
