@@ -32,7 +32,7 @@ final vmArgs = Platform.executableArguments;
 /// generated files around after a failed test. For debugging, we omit that
 /// step if [useLocalDirectory] is true. The place we move them to is saved as
 /// [tempDir].
-String get tempDir => _tempDir == null ? _tempDir = _createTempDir() : _tempDir;
+String? get tempDir => _tempDir == null ? _tempDir = _createTempDir() : _tempDir;
 var _tempDir;
 _createTempDir() => useLocalDirectory
     ? '.'
@@ -44,7 +44,7 @@ var useLocalDirectory = false;
 /// applied to all the arguments of [run]. It will ignore a string that
 /// is an absolute path or begins with "--", because some of the arguments
 /// might be command-line options.
-String asTestDirPath([String s]) {
+String? asTestDirPath([String? s]) {
   if (s == null || s.startsWith("--") || path.isAbsolute(s)) return s;
   return path.join(packageDirectory, 'test', 'message_extraction', s);
 }
@@ -53,9 +53,9 @@ String asTestDirPath([String s]) {
 /// applied to all the arguments of [run]. It will ignore a string that
 /// is an absolute path or begins with "--", because some of the arguments
 /// might be command-line options.
-String asTempDirPath([String s]) {
+String? asTempDirPath([String? s]) {
   if (s == null || s.startsWith("--") || path.isAbsolute(s)) return s;
-  return path.join(tempDir, s);
+  return path.join(tempDir!, s);
 }
 
 typedef Future<ProcessResult> ThenResult(ProcessResult _);
@@ -93,9 +93,9 @@ void copyFilesToTempDirectory() {
     '.packages' // Copy this so that package test can find the imports
   ];
   for (var filename in files) {
-    var file = new File(filename);
+    var file = new File(filename!);
     if (file.existsSync()) {
-      file.copySync(path.join(tempDir, path.basename(filename)));
+      file.copySync(path.join(tempDir!, path.basename(filename)));
     }
   }
 }
@@ -103,7 +103,7 @@ void copyFilesToTempDirectory() {
 void deleteGeneratedFiles() {
   if (useLocalDirectory) return;
   try {
-    new Directory(tempDir).deleteSync(recursive: true);
+    new Directory(tempDir!).deleteSync(recursive: true);
   } on Error catch (e) {
     print("Failed to delete $tempDir");
     print("Exception:\n$e");
@@ -114,12 +114,12 @@ void deleteGeneratedFiles() {
 /// are in dir() and need to be qualified in case that's not our working
 /// directory.
 Future<ProcessResult> run(
-    ProcessResult previousResult, List<String> filenames) {
+    ProcessResult? previousResult, List<String?> filenames) {
   // If there's a failure in one of the sub-programs, print its output.
   checkResult(previousResult);
   var filesInTheRightDirectory = filenames
       .map((x) => asTempDirPath(x))
-      .map((x) => path.normalize(x))
+      .map((x) => path.normalize(x!))
       .toList();
   // Inject the script argument --output-dir in between the script and its
   // arguments.
@@ -133,7 +133,7 @@ Future<ProcessResult> run(
   return result;
 }
 
-checkResult(ProcessResult previousResult) {
+checkResult(ProcessResult? previousResult) {
   if (previousResult != null) {
     if (previousResult.exitCode != 0) {
       print("Error running sub-program:");
@@ -146,7 +146,7 @@ checkResult(ProcessResult previousResult) {
   }
 }
 
-Future<ProcessResult> extractMessages(ProcessResult previousResult) =>
+Future<ProcessResult> extractMessages(ProcessResult? previousResult) =>
     run(previousResult, [
       asTestDirPath('../../bin/extract_to_arb.dart'),
       '--suppress-warnings',
